extract chatTextarea from group/tavern - extract staffList array
